+++ PostType/Quote.php
@@ -10,13 +10,13 @@ if ( !defined( 'ABSPATH' ) ) {
 }
 
 /**
- * Testimonial Post Type
+ * Quote Post Type
  *
  * @package        PostType
  * @since          1.0.0
  * @author         Javier Prieto <jprieton@gmail.com>
  */
-class Testimonial {
+class Quote {
 
   /**
    * Class constructor
@@ -30,28 +30,28 @@ class Testimonial {
   }
 
   /**
-   * Register Testimonial Post Type.
+   * Register Quote Post Type.
    *
    * @since 1.0.0
    */
   public function post_type() {
 
     $labels  = array(
-        'name'                  => __( 'Testimonials', \SourceFramework\TEXTDOMAIN ),
-        'singular_name'         => __( 'Testimonial', \SourceFramework\TEXTDOMAIN ),
-        'menu_name'             => __( 'Testimonials', \SourceFramework\TEXTDOMAIN ),
-        'name_admin_bar'        => __( 'Testimonial', \SourceFramework\TEXTDOMAIN ),
+        'name'                  => __( 'Quotes', \SourceFramework\TEXTDOMAIN ),
+        'singular_name'         => __( 'Quote', \SourceFramework\TEXTDOMAIN ),
+        'menu_name'             => __( 'Quotes', \SourceFramework\TEXTDOMAIN ),
+        'name_admin_bar'        => __( 'Quote', \SourceFramework\TEXTDOMAIN ),
         'archives'              => __( 'Item Archives', \SourceFramework\TEXTDOMAIN ),
-        'all_items'             => __( 'All Testimonials', \SourceFramework\TEXTDOMAIN ),
-        'add_new_item'          => __( 'Add New Testimonial', \SourceFramework\TEXTDOMAIN ),
-        'add_new'               => __( 'New Testimonial', \SourceFramework\TEXTDOMAIN ),
+        'all_items'             => __( 'All Quotes', \SourceFramework\TEXTDOMAIN ),
+        'add_new_item'          => __( 'Add New Quote', \SourceFramework\TEXTDOMAIN ),
+        'add_new'               => __( 'New Quote', \SourceFramework\TEXTDOMAIN ),
         'new_item'              => __( 'New Item', \SourceFramework\TEXTDOMAIN ),
-        'edit_item'             => __( 'Edit Testimonial', \SourceFramework\TEXTDOMAIN ),
-        'update_item'           => __( 'Update Testimonial', \SourceFramework\TEXTDOMAIN ),
-        'view_item'             => __( 'View Testimonial', \SourceFramework\TEXTDOMAIN ),
-        'search_items'          => __( 'Search testimonial', \SourceFramework\TEXTDOMAIN ),
-        'not_found'             => __( 'No testimonial found', \SourceFramework\TEXTDOMAIN ),
-        'not_found_in_trash'    => __( 'No testimonial found in Trash', \SourceFramework\TEXTDOMAIN ),
+        'edit_item'             => __( 'Edit Quote', \SourceFramework\TEXTDOMAIN ),
+        'update_item'           => __( 'Update Quote', \SourceFramework\TEXTDOMAIN ),
+        'view_item'             => __( 'View Quote', \SourceFramework\TEXTDOMAIN ),
+        'search_items'          => __( 'Search quote', \SourceFramework\TEXTDOMAIN ),
+        'not_found'             => __( 'No quote found', \SourceFramework\TEXTDOMAIN ),
+        'not_found_in_trash'    => __( 'No quote found in Trash', \SourceFramework\TEXTDOMAIN ),
         'insert_into_item'      => __( 'Insert into item', \SourceFramework\TEXTDOMAIN ),
         'uploaded_to_this_item' => __( 'Uploaded to this item', \SourceFramework\TEXTDOMAIN ),
         'items_list'            => __( 'Items list', \SourceFramework\TEXTDOMAIN ),
@@ -59,64 +59,61 @@ class Testimonial {
         'filter_items_list'     => __( 'Filter items list', \SourceFramework\TEXTDOMAIN ),
     );
     $rewrite = array(
-        'slug'       => _x( 'testimonial', 'post_type slug', \SourceFramework\TEXTDOMAIN ),
+        'slug'       => _x( 'quote', 'post_type slug', \SourceFramework\TEXTDOMAIN ),
         'with_front' => true,
         'pages'      => true,
         'feeds'      => true,
     );
     $args    = array(
-        'label'               => __( 'Testimonial', \SourceFramework\TEXTDOMAIN ),
+        'label'               => __( 'Quote', \SourceFramework\TEXTDOMAIN ),
         'labels'              => $labels,
         'supports'            => array( 'title', 'editor', 'excerpt', 'thumbnail', ),
-        'taxonomies'          => array( 'testimonial_cat', 'testimonial_tag' ),
+        'taxonomies'          => array( 'quote_cat', 'quote_tag' ),
         'hierarchical'        => false,
         'public'              => true,
         'show_ui'             => true,
         'show_in_menu'        => true,
         'menu_position'       => 20,
-        'menu_icon'           => 'dashicons-testimonial',
+        'menu_icon'           => 'dashicons-format-quote',
         'show_in_admin_bar'   => true,
         'show_in_nav_menus'   => true,
         'can_export'          => true,
-        'has_archive'         => _x( 'testimonial', 'post_type archive_slug', \SourceFramework\TEXTDOMAIN ),
+        'has_archive'         => _x( 'quotes', 'post_type archive_slug', \SourceFramework\TEXTDOMAIN ),
         'exclude_from_search' => false,
         'publicly_queryable'  => true,
         'rewrite'             => $rewrite,
         'capability_type'     => 'post',
     );
-    register_post_type( 'testimonial', $args );
+    register_post_type( 'quote', $args );
   }
 
   /**
-   * Register Testimonial Category Taxonomy.
+   * Register Quote Category Taxonomy.
    *
    * @since 1.0.0
    */
   public function category() {
     $labels  = array(
-        'name'                       => __( 'Testimonial Categories', \SourceFramework\TEXTDOMAIN ),
-        'singular_name'              => __( 'Testimonial Category', \SourceFramework\TEXTDOMAIN ),
+        'name'                       => __( 'Quote Categories', \SourceFramework\TEXTDOMAIN ),
+        'singular_name'              => __( 'Quote Category', \SourceFramework\TEXTDOMAIN ),
         'menu_name'                  => __( 'Categories', \SourceFramework\TEXTDOMAIN ),
-        'all_items'                  => __( 'All Testimonial Categories', \SourceFramework\TEXTDOMAIN ),
-        'parent_item'                => __( 'Parent Testimonial Categories', \SourceFramework\TEXTDOMAIN ),
-        'parent_item_colon'          => __( 'Parent Testimonial Categories:', \SourceFramework\TEXTDOMAIN ),
-        'new_item_name'              => __( 'New Testimonial Categories Name', \SourceFramework\TEXTDOMAIN ),
+        'parent_item'                => __( 'Parent Quote Categories', \SourceFramework\TEXTDOMAIN ),
+        'parent_item_colon'          => __( 'Parent Quote Categories:', \SourceFramework\TEXTDOMAIN ),
+        'new_item_name'              => __( 'New Quote Categories Name', \SourceFramework\TEXTDOMAIN ),
         'add_new_item'               => __( 'Add New Item', \SourceFramework\TEXTDOMAIN ),
-        'edit_item'                  => __( 'Edit Testimonial Categories', \SourceFramework\TEXTDOMAIN ),
-        'update_item'                => __( 'Update Testimonial Categories', \SourceFramework\TEXTDOMAIN ),
+        'edit_item'                  => __( 'Edit Quote Categories', \SourceFramework\TEXTDOMAIN ),
+        'update_item'                => __( 'Update Quote Categories', \SourceFramework\TEXTDOMAIN ),
         'view_item'                  => __( 'View Item', \SourceFramework\TEXTDOMAIN ),
-        'separate_items_with_commas' => __( 'Separate items with commas', \SourceFramework\TEXTDOMAIN ),
         'add_or_remove_items'        => __( 'Add or remove items', \SourceFramework\TEXTDOMAIN ),
-        'choose_from_most_used'      => __( 'Choose from the most used', \SourceFramework\TEXTDOMAIN ),
         'popular_items'              => __( 'Popular Items', \SourceFramework\TEXTDOMAIN ),
-        'search_items'               => __( 'Search Testimonial Categories', \SourceFramework\TEXTDOMAIN ),
+        'search_items'               => __( 'Search Quote Categories', \SourceFramework\TEXTDOMAIN ),
         'not_found'                  => __( 'Not Found', \SourceFramework\TEXTDOMAIN ),
         'no_terms'                   => __( 'No items', \SourceFramework\TEXTDOMAIN ),
         'items_list'                 => __( 'Items list', \SourceFramework\TEXTDOMAIN ),
         'items_list_navigation'      => __( 'Items list navigation', \SourceFramework\TEXTDOMAIN ),
     );
     $rewrite = array(
-        'slug'         => _x( 'testimonial-category', 'taxonomy slug', \SourceFramework\TEXTDOMAIN ),
+        'slug'         => _x( 'quote-category', 'taxonomy slug', \SourceFramework\TEXTDOMAIN ),
         'with_front'   => false,
         'hierarchical' => true,
     );
@@ -130,39 +127,35 @@ class Testimonial {
         'show_tagcloud'     => true,
         'rewrite'           => $rewrite,
     );
-    register_taxonomy( 'testimonial_cat', array( 'testimonial' ), $args );
+    register_taxonomy( 'quote_cat', array( 'quote' ), $args );
   }
 
   /**
-   * Register Testimonial Tag Taxonomy.
+   * Register Quote Tag Taxonomy.
    *
    * @since 1.0.0
    */
   public function tag() {
     $labels  = array(
-        'name'                       => __( 'Testimonial Tags', \SourceFramework\TEXTDOMAIN ),
-        'singular_name'              => __( 'Testimonial Tag', \SourceFramework\TEXTDOMAIN ),
+        'name'                       => __( 'Quote Tags', \SourceFramework\TEXTDOMAIN ),
+        'singular_name'              => __( 'Quote Tag', \SourceFramework\TEXTDOMAIN ),
         'menu_name'                  => __( 'Tags', \SourceFramework\TEXTDOMAIN ),
-        'all_items'                  => __( 'All Testimonial Tags', \SourceFramework\TEXTDOMAIN ),
-        'parent_item'                => __( 'Parent Testimonial Tags', \SourceFramework\TEXTDOMAIN ),
-        'parent_item_colon'          => __( 'Parent Testimonial Tags:', \SourceFramework\TEXTDOMAIN ),
-        'new_item_name'              => __( 'New Testimonial Tags Name', \SourceFramework\TEXTDOMAIN ),
+        'parent_item'                => __( 'Parent Quote Tags', \SourceFramework\TEXTDOMAIN ),
+        'parent_item_colon'          => __( 'Parent Quote Tags:', \SourceFramework\TEXTDOMAIN ),
+        'new_item_name'              => __( 'New Quote Tags Name', \SourceFramework\TEXTDOMAIN ),
         'add_new_item'               => __( 'Add New Item', \SourceFramework\TEXTDOMAIN ),
-        'edit_item'                  => __( 'Edit Testimonial Tags', \SourceFramework\TEXTDOMAIN ),
-        'update_item'                => __( 'Update Testimonial Tags', \SourceFramework\TEXTDOMAIN ),
+        'edit_item'                  => __( 'Edit Quote Tags', \SourceFramework\TEXTDOMAIN ),
+        'update_item'                => __( 'Update Quote Tags', \SourceFramework\TEXTDOMAIN ),
         'view_item'                  => __( 'View Item', \SourceFramework\TEXTDOMAIN ),
-        'separate_items_with_commas' => __( 'Separate items with commas', \SourceFramework\TEXTDOMAIN ),
         'add_or_remove_items'        => __( 'Add or remove items', \SourceFramework\TEXTDOMAIN ),
-        'choose_from_most_used'      => __( 'Choose from the most used', \SourceFramework\TEXTDOMAIN ),
         'popular_items'              => __( 'Popular Items', \SourceFramework\TEXTDOMAIN ),
-        'search_items'               => __( 'Search Testimonial Tags', \SourceFramework\TEXTDOMAIN ),
-        'not_found'                  => __( 'Not Found', \SourceFramework\TEXTDOMAIN ),
+        'search_items'               => __( 'Search Quote Tags', \SourceFramework\TEXTDOMAIN ),
         'no_terms'                   => __( 'No items', \SourceFramework\TEXTDOMAIN ),
         'items_list'                 => __( 'Items list', \SourceFramework\TEXTDOMAIN ),
         'items_list_navigation'      => __( 'Items list navigation', \SourceFramework\TEXTDOMAIN ),
     );
     $rewrite = array(
-        'slug'         => _x( 'testimonial-tag', 'taxonomy slug', \SourceFramework\TEXTDOMAIN ),
+        'slug'         => _x( 'quote-tag', 'taxonomy slug', \SourceFramework\TEXTDOMAIN ),
         'with_front'   => false,
         'hierarchical' => true,
     );
@@ -176,7 +169,7 @@ class Testimonial {
         'show_tagcloud'     => true,
         'rewrite'           => $rewrite,
     );
-    register_taxonomy( 'testimonial_tag', array( 'testimonial' ), $args );
+    register_taxonomy( 'quote_tag', array( 'quote' ), $args );
   }
 
 }