+++ CHANGELOG.md
@@ -4,6 +4,16 @@ Starting with version 1.1, the SourceFramework follows [SemVer](http://semver.or
 
 ## 1.x
 
+- 1.3.0
+  - Change some strings to default translations
+  - Post Formats
+  - Quote post type 
+  - Option to enable shortcodes in text widgets
+  - SVG Icons to PNG for use in email templates
+
+- 1.2.0
+  - Enables the Excerpt meta box in Page edit screen.
+
 - 1.1.0
   - Bootstrap 3 & 4 Breadcrumb
   - Optional thumbnail column for post types
@@ -11,6 +21,3 @@ Starting with version 1.1, the SourceFramework follows [SemVer](http://semver.or
   - reCAPTCHA API
   - Input, User and Error classes
   - Added FrontEnd Helper compatibility with Boostrap 4.x 
-
-- 1.2.0
-  - Enables the Excerpt meta box in Page edit screen.